+++ praw/models/reddit/submission/__init__.py
@@ -1,0 +1,5 @@
+"""Container for Submission and related classes."""
+# flake8: noqa
+from .flair import SubmissionFlair
+from .moderation import SubmissionModeration
+from .submission import Submission
